Add annotation campaign and rework segment collection
This allows us to track a per-user set of segments and per-user progress in that set.

The segment collection has also been changed from an async iterator to more of an arraylist-like interface.
This interface is far simpler and allows us to more efficiently prefetch segments in the background without holding up the current segment.

Also fixes a couple of race conditions encountered when rapidly advancing through segments.